requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
python-dotenv==1.0.1
flask-mail==0.10.0 